Struct bholdus_dex::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn swap_with_exact_supply(
origin: OriginFor<T>,
path: Vec<CurrencyId>,
supply_amount: Balance,
min_target_amount: Balance
) -> DispatchResult

Trading with DEX, swap with exact supply amount.

path: trading path.

supply_amount: exact supply amount.

min_target_amount: acceptable target amount.

TODO: Weight benchmarking

pub fn swap_with_exact_target(
origin: OriginFor<T>,
path: Vec<CurrencyId>,
target_amount: Balance,
max_supply_amount: Balance
) -> DispatchResult

Trading with DEX, swap with exact target amount.

path: trading path.

target_amount: target amount user wants to receive.

max_supply_amount: maximum supply amount that user willing to pay.

Weight benchmarking

pub fn start_trading_pair_provisioning(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
min_contribution_0: Balance,
min_contribution_1: Balance,
target_contribution_0: Balance,
target_contribution_1: Balance
) -> DispatchResult

Start provisioning stage of a trading pair

TODO: Weight benchmarking

pub fn update_trading_pair_provisioning_parameters(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
min_contribution_0: Balance,
min_contribution_1: Balance,
target_contribution_0: Balance,
target_contribution_1: Balance
) -> DispatchResult

Update provisioning parameters of provisioning trading pair

TODO: Weight benchmarking

pub fn add_trading_pair_provision(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
contribution_0: Balance,
contribution_1: Balance
) -> DispatchResult

Add provision to a provisioning trading pair

TODO: Weight benchmarking

pub fn enable_provisioning_trading_pair(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId
) -> DispatchResult

Enable provisioning trading pair

TODO: Weight benchmarking

pub fn enable_trading_pair(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId
) -> DispatchResult

Enable trading pair from disabled or provisioning with zero accumulated provision

TODO: Weight benchmarking

pub fn add_liquidity(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
max_amount_0: Balance,
max_amount_1: Balance
) -> DispatchResult

Add liquidity to enabled trading pair

TODO: Weight benchmarking

pub fn remove_liquidity(
origin: OriginFor<T>,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
remove_share_amount: Balance,
min_withdrawn_amount_0: Balance,
min_withdrawn_amount_1: Balance
) -> DispatchResult

Remove liquidity

TODO: Weight benchmarking

pub fn claim_dex_share(
origin: OriginFor<T>,
owner: T::AccountId,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId
) -> DispatchResult

Claim dex share. Founders that add provision to trading pair in provisioning stage can use this call to claim their dex share

impl<T: Config> Pallet<T>

pub fn liquidity_pool<KArg>(k: KArg) -> (Balance, Balance) where
KArg: EncodeLike<TradingPair>,

Reserves of trading pairs.

impl<T: Config> Pallet<T>

pub fn trading_pair_statuses<KArg>(k: KArg) -> TradingPairStatus<Balance> where
KArg: EncodeLike<TradingPair>,

Status of trading pairs.

impl<T: Config> Pallet<T>

pub fn provisioning_pool<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> (Balance, Balance) where
KArg1: EncodeLike<TradingPair>,
KArg2: EncodeLike<T::AccountId>,

Provision of each user added to a trading pair when that trading pair is in provisioning status.

impl<T: Config> Pallet<T>

pub fn initial_share_exchange_rates<KArg>(
k: KArg
) -> (ExchangeRate, ExchangeRate) where
KArg: EncodeLike<TradingPair>,

Initial share exchange rates for each trading pair. Used to calculate share amount for first liqudity providers.

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

pub fn get_share_and_pool_increment(
pool_0: Balance,
pool_1: Balance,
max_amount_0: Balance,
max_amount_1: Balance,
total_share_amount: Balance
) -> Result<(Balance, Balance, Balance), DispatchError>

Get share and pool increment

pool_0: current liquidity pool of currency_0

pool_1: current liquidity pool of currency_1

max_amount_0: maximum amount of currency_0 users wants to add liquidity

max_amount_1: maximum amount of currency_1 users wants to add liquidity

total_share_amount: current total shares of this trading pair.

total_share_amount = 0 indicates that calculation for first liquidity providers should be taken. In this case, pool_0 and pool_1 is ignored.

Returns: (pool_0_increment, pool_1_increment, share_increment)

pub fn get_share_and_pool_decrement(
pool_0: Balance,
pool_1: Balance,
remove_share_amount: Balance,
total_share_amount: Balance
) -> Result<(Balance, Balance, Balance), DispatchError>

Return share and pool decrement when users withdraw liquidity token

pool_0: liquidity pool of currency_0

pool_1: liqudity pool of currency_1

remove_share_amount: share amount to remove

total_share_amount: total share amount

Returns: (pool_0_decrement, pool_1_decrement, share_decrement)

pub fn do_claim_dex_share(
owner: &T::AccountId,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId
) -> DispatchResult

pub fn do_remove_liquidity(
who: &T::AccountId,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
remove_share_amount: Balance,
min_withdrawn_amount_0: Balance,
min_withdrawn_amount_1: Balance
) -> DispatchResult

pub fn do_add_liquidity(
who: &T::AccountId,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
max_amount_0: Balance,
max_amount_1: Balance
) -> DispatchResult

pub fn do_enable_trading_pair(
currency_id_0: CurrencyId,
currency_id_1: CurrencyId
) -> DispatchResult

pub fn do_enable_provisioning_trading_pair(
currency_id_0: CurrencyId,
currency_id_1: CurrencyId
) -> DispatchResult

pub fn do_add_trading_pair_provision(
who: &T::AccountId,
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
contribution_0: Balance,
contribution_1: Balance
) -> DispatchResult

pub fn do_update_trading_pair_provisioning_parameters(
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
min_contribution_0: Balance,
min_contribution_1: Balance,
target_contribution_0: Balance,
target_contribution_1: Balance
) -> DispatchResult

pub fn do_start_trading_pair_provisioning(
currency_id_0: CurrencyId,
currency_id_1: CurrencyId,
min_contribution_0: Balance,
min_contribution_1: Balance,
target_contribution_0: Balance,
target_contribution_1: Balance
) -> DispatchResult

Implementation of start_trading_pair_provisioning

pub fn do_swap_with_exact_supply(
who: &T::AccountId,
path: &[CurrencyId],
supply_amount: Balance,
min_target_amount: Balance
) -> Result<Balance, DispatchError>

Implementation of swap_with_exact_supply

Returns a Result:

target_amount: calculated target amount when successful

error: dispatch error when failed

pub fn do_swap_with_exact_target(
who: &T::AccountId,
path: &[CurrencyId],
target_amount: Balance,
max_supply_amount: Balance
) -> Result<Balance, DispatchError>

Implementation of swap_with_exact_target

Returns a Result:

supply_amount: Calculated amount that user must pay

error: dispatch error when failed

pub fn get_liquidity(
currency_id_0: &CurrencyId,
currency_id_1: &CurrencyId
) -> (Balance, Balance)

Returns reserves of two currencies.

pub fn get_target_amount(
supply_pool: Balance,
target_pool: Balance,
supply_amount: Balance
) -> Balance

Returns target amount, given supply reserves, target reserves, supply amount.

pub fn get_target_amounts(
path: &[CurrencyId],
supply_amount: Balance
) -> Result<Vec<Balance>, DispatchError>

Returns list of target amounts with each is the target amount of a trading pair of given trading path.

pub fn get_supply_amount(
supply_pool: Balance,
target_pool: Balance,
target_amount: Balance
) -> Balance

pub fn get_supply_amounts(
path: &[CurrencyId],
target_amount: Balance
) -> Result<Vec<Balance>, DispatchError>