Trait bholdus_dex::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type ListingOrigin: EnsureOrigin<Self::Origin>;
    type PalletId: Get<PalletId>;
    type ExchangeFee: Get<(u32, u32)>;
    type TradingPathLimit: Get<u32>;
    type Currency: MultiCurrency<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type ListingOrigin: EnsureOrigin<Self::Origin>

The origin that may disable, enable, listing provisioning trading pair

type PalletId: Get<PalletId>

Pallet Id turned into Account Id to hold tokens

type ExchangeFee: Get<(u32, u32)>

Exchange fee. Use seperate numerator, denominator to achieve result more accurate

type TradingPathLimit: Get<u32>

Trading path limit

type Currency: MultiCurrency<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance>

Multi currency mechanism

Loading content...